requirements.txt
ezodf>=0.3.2
gspread>=5.7.2
oauth2client>=4.1.3
PyYAML>=6.0
XlsxWriter>=3.0.8
